Revert "main: watch configuration directory for changes"
The idea of having a watcher checking the configuration directory for
changes is interesting in principle. One could also extend it so that
destination files are watched too, in order to allow no divergence
between desired state and reality.

However, a one-shot approach is much simpler to understand for users, as
well as to implement and debug. In the field of software configuration
management it is probably wise to follow the principle of least
surprise: for example, changes should not get automatically applied as
soon as the configuration is updated, but only upon admins's explicit
initiative.

This reverts commit 9fd09ceeed7c6be816af5a229560cde19cd892ab.