package: fix IsInstalled() logic on Alpine
The implementation of IsInstalled on Alpine always returned true because
apk info -qe $not_installed_pkg exits with 1 but without any error as
far as cmd.Run() is concerned. Luckily apk info -e $installed_pkg prints
the package name to standard output if the package is installed, and
nothing if it is not. Use the fact to implement IsInstalled() on Alpine.