Rename NewPetsFile to NewTestFile
The function NewPetsFile has been used so far only for testing purposes.

Given that we need an actual constructor for the PetsFile struct,
specifically a function that takes no argument and creates a new
PetsFile with default values, rename NewPetsFile to NewTestFile so that
we can later define NewPetsFile as the constructor.